requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-docx>=1.0.0
pdfplumber>=0.10.0
scikit-learn>=1.3.0
pydantic>=2.5.0
python-multipart>=0.0.6
